+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ## [Unreleased]
 
+### Added
+
+- Optional `platforms` configuration field for binaries to restrict them to only selected platforms (either os or os/arch values)
+
+### Fixed
+
+- The asset names for codex
+
 ## [5.3.0] - 2026-08-23
 
 ### Fixed